Move Options to a file. Actually run the extension depending on options
